Move tests to ticket namespace (and rename them)
