+++ tests/Doctrine/Tests/ORM/Functional/Ticket/Ticket4646InstanceOfTest.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Doctrine\Tests\ORM\Functional {
+namespace Doctrine\Tests\ORM\Functional\Ticket {
 
-    use Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest\Employee;
-    use Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest\Person;
+    use Doctrine\Tests\ORM\Functional\InstanceOfTest\Employee;
+    use Doctrine\Tests\ORM\Functional\InstanceOfTest\Person;
     use Doctrine\Tests\OrmFunctionalTestCase;
 
-    class InstanceOfAbstractTest extends OrmFunctionalTestCase
+    class Ticket4646InstanceOfTest extends OrmFunctionalTestCase
     {
         protected function setUp()
         {
             parent::setUp();
 
             $this->_schemaTool->createSchema([
-                $this->_em->getClassMetadata(Person::class),
-                $this->_em->getClassMetadata(Employee::class),
+                $this->_em->getClassMetadata(__NAMESPACE__ . '\InstanceOfTest\Person'),
+                $this->_em->getClassMetadata(__NAMESPACE__ . '\InstanceOfTest\Employee'),
             ]);
         }
 
@@ -22,45 +22,52 @@ namespace Doctrine\Tests\ORM\Functional {
         {
             $this->loadData();
 
-            $dql = 'SELECT p FROM Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest\Person p
-                    WHERE p INSTANCE OF Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest\Person';
+            $dql = 'SELECT p FROM Doctrine\Tests\ORM\Functional\InstanceOfTest\Person p
+                    WHERE p INSTANCE OF Doctrine\Tests\ORM\Functional\InstanceOfTest\Person';
             $query = $this->_em->createQuery($dql);
             $result = $query->getResult();
 
-            $this->assertCount(1, $result);
+            $this->assertCount(2, $result);
 
             foreach ($result as $r) {
-                $this->assertInstanceOf(InstanceOfAbstractTest\Person::class, $r);
-                $this->assertInstanceOf(InstanceOfAbstractTest\Employee::class, $r);
-                $this->assertSame('bar', $r->getName());
+                $this->assertInstanceOf(Person::class, $r);
+                if ($r instanceof Employee) {
+                    $this->assertEquals('bar', $r->getName());
+                } else {
+                    $this->assertEquals('foo', $r->getName());
+                }
             }
         }
 
         private function loadData()
         {
-            $employee = new InstanceOfAbstractTest\Employee();
+            $person = new Person();
+            $person->setName('foo');
+
+            $employee = new Employee();
             $employee->setName('bar');
             $employee->setDepartement('qux');
 
+            $this->_em->persist($person);
             $this->_em->persist($employee);
 
-            $this->_em->flush($employee);
+            $this->_em->flush(array($person, $employee));
         }
     }
 }
 
-namespace Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest {
-
+namespace Doctrine\Tests\ORM\Functional\InstanceOfTest {
     /**
      * @Entity()
-     * @Table(name="instance_of_abstract_test_person")
+     * @Table(name="instance_of_test_person")
      * @InheritanceType(value="JOINED")
      * @DiscriminatorColumn(name="kind", type="string")
      * @DiscriminatorMap(value={
-     *     "employee": Employee::class
+     *     "person": "Doctrine\Tests\ORM\Functional\InstanceOfTest\Person",
+     *     "employee": "Doctrine\Tests\ORM\Functional\InstanceOfTest\Employee"
      * })
      */
-    abstract class Person
+    class Person
     {
         /**
          * @Id()
@@ -92,7 +99,7 @@ namespace Doctrine\Tests\ORM\Functional\InstanceOfAbstractTest {
 
     /**
      * @Entity()
-     * @Table(name="instance_of_abstract_test_employee")
+     * @Table(name="instance_of_test_employee")
      */
     class Employee extends Person
     {
